paho_mqtt/requirements.txt
azure-cosmos
paho-mqtt
python-dotenv